requirements.txt
Django==1.8.3
django-celery==3.1.16
django-contrib-comments==1.6.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
django-extensions==1.5.5
django-fluent-comments==0.9.2
django-grappelli==2.5.3
sorl-thumbnail==12.3
django-threadedcomments==0.9.0
timezones
python-social-auth
django-tags-input
git+git://github.com/gkuhn1/django-admin-multiupload.git
